+++ src/parsers/problem/CodeforcesProblemParser.ts
@@ -8,10 +8,15 @@ export class CodeforcesProblemParser extends Parser {
   getMatchPatterns(): string[] {
     return [
       'http://codeforces.com/contest/*/problem/*',
+      'https://codeforces.com/contest/*/problem/*',
       'http://codeforces.com/problemset/problem/*/*',
+      'https://codeforces.com/problemset/problem/*/*',
       'http://codeforces.com/gym/*/problem/*',
+      'https://codeforces.com/gym/*/problem/*',
       'http://codeforces.com/group/*/contest/*/problem/*',
+      'https://codeforces.com/group/*/contest/*/problem/*',
       'http://codeforces.com/problemsets/acmsguru/problem/*/*',
+      'https://codeforces.com/problemsets/acmsguru/problem/*/*',
     ];
   }