+++ tests/parser-functions.ts
@@ -19,6 +19,7 @@ export default {
   },
 
   async beforeHackerRank(page: Page, data: ParserTestData) {
+    await page.waitFor('.login');
     await page.click('.login');
     await page.type('#legacy-login input[name=login]', process.env.HACKER_RANK_EMAIL);
     await page.type('#legacy-login input[name=password]', process.env.HACKER_RANK_PASSWORD);